+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Static attributes are now denoted with a leading single quote, such as `'msg Hello!` rather than a double quote.
 - Self-closing elements now have proper trailing newline when `keepWhitespace` is enabled.
 - Unanchored attributes (attributes whose indentation don't match a preceding tag) are now properly discarded.
+- Non-Louk template section languages are now properly respected.
 
 1.0.2
 ------------------