+++ CHANGELOG.md
@@ -1,9 +1,16 @@
+2.0.2
+------------------
+- Unanchored attributes (attributes whose indentation don't match a preceding tag) are now properly discarded.
+
+2.0.1
+------------------
+- Documentation improvements
+
 2.0.0
 ------------------
 - Multiline content is supported by beginning lines with `|`.
 - Static attributes are now denoted with a leading single quote, such as `'msg Hello!` rather than a double quote.
 - Self-closing elements now have proper trailing newline when `keepWhitespace` is enabled.
-- Unanchored attributes (attributes whose indentation don't match a preceding tag) are now properly discarded.
 - Non-Louk template section languages are now properly respected.
 
 1.0.2